Exceptionless telemeters are added

+++ TIKSN.Framework.Shared/Telemetry/ApplicationInsightsMetricTelemeter.cs
@@ -5,9 +5,9 @@ namespace TIKSN.Analytics.Telemetry
 {
 	public class ApplicationInsightsMetricTelemeter : IMetricTelemeter
 	{
-		public async Task TrackMetric(string metricName, double metricValue)
+		public async Task TrackMetric(string metricName, decimal metricValue)
 		{
-			var telemetry = new MetricTelemetry(metricName, metricValue);
+			var telemetry = new MetricTelemetry(metricName, (double)metricValue);
 
 			ApplicationInsightsHelper.TrackMetric(telemetry);
 		}
